requirements.txt
appdata==2.1.2
cachelib==0.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.12
click==8.0.3
colorama==0.4.5
colored==1.4.3
dataclasses==0.8
Flask==2.0.3
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.2
Flask-Session==0.4.0
idna==3.4
importlib-metadata==4.8.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
paypal==1.2.5
pyasn1==0.4.8
pycparser==2.21
PyJWT==2.4.0
pymongo==4.1.1
pyotp==2.7.0
python-dateutil==2.8.2
python-dotenv==0.20.0
requests==2.27.1
rsa==4.9
six==1.16.0
to-requirements.txt==1.3.0
typing-extensions==4.1.1
urllib3==1.26.15
Werkzeug==2.0.3
zipp==3.6.0
